Rename resource to http
