+++ wazo_webhookd/plugins/api/plugin.py
@@ -1,7 +1,7 @@
-# Copyright (C) 2015-2016 Avencall
+# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resources import SwaggerResource
+from .http import SwaggerResource
 
 
 class Plugin(object):
