+++ wazo_webhookd/plugins/config/plugin.py
@@ -1,7 +1,7 @@
-# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resources import ConfigResource
+from .http import ConfigResource
 
 
 class Plugin(object):
