+++ wazo_webhookd/plugins/status/plugin.py
@@ -1,7 +1,7 @@
-# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resource import StatusResource
+from .http import StatusResource
 
 
 class Plugin(object):
